# contracts/platform-api.openapi.yaml
openapi: 3.1.0
info:
  title: 台区能源辨识与光伏功率分离平台 API
  version: 2.0.0
  description: React 前端仅调用本 Java API。系统只提供监测分析，不提供设备控制。
servers:
  - url: /api/v1
security:
  - oidcBearer: []
paths:
...
  /stations/{stationId}/inference-replays:
    get:
      operationId: listInferenceReplayJobs
      parameters: [{ $ref: '#/components/parameters/StationId' }]
      responses:
        '200':
          description: 最近历史推理回放任务
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/InferenceReplayJob' } }
    post:
      operationId: createInferenceReplayJob
      description: 对数据库中已存在的总开分钟数据执行真实模型历史补算；不会伪造缺失结果。
      parameters: [{ $ref: '#/components/parameters/StationId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InferenceReplayRequest' }
      responses:
        '202':
          description: 回放任务已进入队列
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InferenceReplayJob' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /stations/{stationId}/inference-replays/{jobId}:
    get:
      operationId: getInferenceReplayJob
      parameters:
        - $ref: '#/components/parameters/StationId'
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 历史推理回放任务状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InferenceReplayJob' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /imports:
    post:
      operationId: createImport
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [station_id, data_type, file]
              properties:
                station_id: { type: string }
                data_type: { type: string, enum: [main_switch, pv_feedback] }
                file: { type: string, format: binary }
      responses:
        '202':
          description: 已接受导入
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportResult' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /exports:
    post:
      operationId: createExport
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExportRequest' }
      responses:
        '200':
          description: CSV 导出
          content:
            text/csv:
              schema: { type: string, format: binary }
  /ingestion/main-switch/minutes:
    post:
      operationId: ingestMainSwitchMinutes
      parameters:
        - in: query
          name: replace_existing
          description: 仅用于受控数据修复；字段不同的同站同分钟记录会被更新。
          schema: { type: boolean, default: false }
        - in: query
          name: run_inference
          description: 批量历史导入时设为 false，完成后通过历史回放统一推理。
          schema: { type: boolean, default: true }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [points]
              properties:
                points:
                  type: array
                  minItems: 1
                  maxItems: 5000
                  items: { $ref: '#/components/schemas/MainSwitchMinutePoint' }
      responses:
        '202':
          description: 原始分钟已幂等保存；是否进入逐分钟推理由 run_inference 决定
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestionReceipt' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /ingestion/pv-feedback/batches:
    post:
      operationId: ingestPvFeedbackBatch
      description: 反馈仅在 arrival_time 到达后按历史 event_time 生成校正。
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '202': { description: 反馈批次已幂等接收 }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /models/{version}/approve:
    post:
      operationId: approveModel
      parameters: [{ name: version, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: 模型已人工审批 }
  /models/{version}/deploy:
    post:
      operationId: deployModel
      parameters: [{ name: version, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: 模型已部署为 active 或 candidate }
  /models/{task}/rollback:
    post:
      operationId: rollbackModel
      parameters: [{ name: task, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: 已回滚到上一 active 版本 }
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    StationId:
      name: stationId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 64 }
...
  responses:
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: 数据校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    InferenceReplayRequest:
      type: object
      additionalProperties: false
      required: [from, to]
      properties:
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        dry_run: { type: boolean, default: false }
    InferenceReplayJob:
      type: object
      required: [job_id, station_id, range_start, range_end, status, dry_run, total_targets,
        processed_targets, recognition_results, separation_results, skipped_targets, failed_targets,
        request_id, requested_by, created_at, updated_at]
      properties:
        job_id: { type: string }
        station_id: { type: string }
        range_start: { type: string, format: date-time }
        range_end: { type: string, format: date-time }
        status: { type: string, enum: [queued, running, completed, failed] }
        dry_run: { type: boolean }
        total_targets: { type: integer, minimum: 0 }
        processed_targets: { type: integer, minimum: 0 }
        recognition_results: { type: integer, minimum: 0 }
        separation_results: { type: integer, minimum: 0 }
        skipped_targets: { type: integer, minimum: 0 }
        failed_targets: { type: integer, minimum: 0 }
        error_summary: { type: [string, 'null'] }
        request_id: { type: string }
        requested_by: { type: string }
        created_at: { type: string, format: date-time }
        started_at: { type: [string, 'null'], format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        updated_at: { type: string, format: date-time }
    MainSwitchMinutePoint:
      type: object
      additionalProperties: false
      required: [station_id, event_time, active_power_kw, phase_a_power_kw, phase_b_power_kw, phase_c_power_kw, coverage_ratio, quality_flag, source_id]
      properties:
        station_id: { type: string }
        event_time: { type: string, format: date-time }
        active_power_kw: { type: number }
        phase_a_power_kw: { type: number }
        phase_b_power_kw: { type: number }
        phase_c_power_kw: { type: number }
        reactive_power_kvar: { type: [number, 'null'] }
        voltage: { type: [number, 'null'] }
        current: { type: [number, 'null'] }
        pf: { type: [number, 'null'] }
        coverage_ratio: { type: number, minimum: 0, maximum: 1 }
        quality_flag: { type: string, enum: [good, warning, missing, out_of_order] }
        source_id: { type: string }
...
    ErrorResponse:
      type: object
      required: [code, message, request_id, details]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        details: { type: object }
    IngestionReceipt:
      type: object
      required: [request_id, accepted, inserted, updated, duplicates, out_of_order, inference_scheduled]
      properties:
        request_id: { type: string }
        accepted: { type: integer }
        inserted: { type: integer }
        updated: { type: integer }
        duplicates: { type: integer }
        out_of_order: { type: integer }
        inference_scheduled: { type: boolean }
    ImportResult:
      type: object
      required: [import_id, status, accepted_rows]
      properties:
        import_id: { type: string }
        status: { type: string, enum: [accepted] }
        accepted_rows: { type: integer, minimum: 0 }
    ExportRequest:
      type: object
      required: [station_id, from, to, format, include]
      properties:
        station_id: { type: string }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        format: { type: string, enum: [csv] }
        include: { type: array, items: { type: string, enum: [recognition_results, separation_results, feedback_batches, corrections] } }
...